In AVLNode, make sumOfDescendantValues int64_t instead of int
The type of sumOfDescendantValues should be at least as expressive as
the type of value, which is int64_t.

+++ my-own-problems/shared/avl-tree.h
@@ -9,7 +9,7 @@ struct AVLNode
     int balanceFactor = 0;
     int maxDescendantDepth = 0;
     int numDescendants = 1;
-    int sumOfDescendantValues = 0;
+    int64_t sumOfDescendantValues = 0;
 
     int id = -1;
 };
